remove need for --sender
